feat: Add new columns for motoristas page
This commit adds new columns to the `columns_motoristas` array in the `columns.tsx` file. The new columns include 'id', 'email', 'seller_code', 'schedule', and 'car'. These columns will be used in the `Table` component to display additional information for the drivers. This change improves the functionality and usability of the motoristas page.