feat: Format schedule date and time in Table component
This commit adds code to format the schedule date and time in the Table component of the motoristas page. The code checks if the column field is "schedule" and converts the date string to the format "dd/mm/yyyy hh:mm". This change improves the readability and usability of the table data.