Refactor CarsContext and Carros page, add update modal
The code changes refactor the CarsContext and Carros page. In the CarsContext, the variable name "cadasteredCars" is corrected to "cadasterCars". In the Carros page, a new state variable "showUpdateModal" is added to control the visibility of the update modal. Additionally, a new state variable "carToUpdate" is added to store the car object that is being updated. The "editAction" function is modified to set the "carToUpdate" state and show the update modal when the edit button is clicked. The updateCar function is added to update the name of the car in the "carToUpdate" state. The update modal is rendered when "showUpdateModal" is true, allowing the user to update the name of the car. This change improves the functionality and usability of the Carros page.