feat: Add Admin interface and related components
The code changes add a new interface called AdminProps, which defines the properties of an admin. This interface includes fields such as id, name, cpf, email, and phone. Additionally, new components and styles are added for the "novo" and "editar" pages of the "administradores" section. The "admins_mock" array is also created to store mock data for admins. These changes enhance the functionality and user experience of the application.